Switch to libfmt 10.0.0

+++ src/dwarfs/compression/lzma.cpp
@@ -152,7 +152,8 @@ lzma_block_compressor::compress(const std::vector<uint8_t>& data,
     if (auto it = lzma_error_desc.find(ret); it != lzma_error_desc.end()) {
       DWARFS_THROW(runtime_error, fmt::format("LZMA error: {}", it->second));
     } else {
-      DWARFS_THROW(runtime_error, fmt::format("LZMA: unknown error {}", ret));
+      DWARFS_THROW(runtime_error, fmt::format("LZMA: unknown error {}",
+                                              static_cast<int>(ret)));
     }
   }
 
@@ -234,7 +235,8 @@ class lzma_block_decompressor final : public block_decompressor::impl {
       if (it != lzma_error_desc.end()) {
         error_ = fmt::format("LZMA: decompression failed ({})", it->second);
       } else {
-        error_ = fmt::format("LZMA: decompression failed (error {})", ret);
+        error_ = fmt::format("LZMA: decompression failed (error {})",
+                             static_cast<int>(ret));
       }
       DWARFS_THROW(runtime_error, error_);
     }
@@ -295,8 +297,8 @@ size_t lzma_block_decompressor::get_uncompressed_size(const uint8_t* data,
 
   lzma_ret ret = lzma_code(&s, LZMA_RUN);
   if (ret != LZMA_STREAM_END || s.avail_in != 0) {
-    DWARFS_THROW(runtime_error,
-                 fmt::format("lzma_code(): {} (avail_in={})", ret, s.avail_in));
+    DWARFS_THROW(runtime_error, fmt::format("lzma_code(): {} (avail_in={})",
+                                            static_cast<int>(ret), s.avail_in));
   }
 
   pos -= LZMA_STREAM_HEADER_SIZE;
